requirements.txt
psycopg2-binary
fastapi
uvicorn
sqlalchemy
pytz
pandas
python-dotenv